Added Button to stop motors
PA0 will now stop the motors, but will allow us to continue reading information from the Gyro.  Once the button is pressed again it should zero out the values it once had and start from 0 on the gyro.   Need a PI or PID badly for this, it is very sensitive to movement.